@@ -1,4 +1,4 @@
-import glob
+import importlib
 import os.path as osp
 from typing import Optional, Tuple
 
@@ -6,8 +6,8 @@ import torch
 
 from .utils import broadcast
 
-torch.ops.load_library(
-    glob.glob(osp.join(osp.dirname(__file__), '_scatter.*'))[0])
+torch.ops.load_library(importlib.machinery.PathFinder().find_spec(
+    '_scatter', [osp.dirname(__file__)]).origin)
 
 
 @torch.jit.script
