@@ -1,11 +1,11 @@
-import glob
+import importlib
 import os.path as osp
 from typing import Optional, Tuple
 
 import torch
 
-torch.ops.load_library(
-    glob.glob(osp.join(osp.dirname(__file__), '_segment_csr.*'))[0])
+torch.ops.load_library(importlib.machinery.PathFinder().find_spec(
+    '_segment_csr', [osp.dirname(__file__)]).origin)
 
 
 @torch.jit.script
